+++ changelog.md
@@ -1,6 +1,9 @@
 # Changelog
 ## Version v2.4.0
 * Added support for Geode v5.0.0 / GD 2.2081.
+* Added master enable and master volume.
+* Added filtering by featured click packs.
+* The ID of a click pack is now shown in its more info page.
 ## Version v2.3.3
 * Fixed a bug on windows where release sounds will continue playing while disabled if the 'sounds everywhere' setting is enabled.
 ## Version v2.3.2